M26: remove the status bar — fold its controls into the Command Window
The status bar is gone entirely (deleted chrome_statusbar.go and its DrawChrome call).
Its remaining interactive state — the running tool's OK/Cancel, the selection count,
and the progress bar — now render as a control row inside the Command Window, shown only
when something is live (a tool, progress, or a selection). The default dock layout docks
the Command Window into the bottom band where the status bar used to sit.

The Command Window is now the single bottom panel: rolling feedback + prompt, the tool
control row, autocomplete, and the input line — no separate Status window.
